Refactored food code into a single route file.
