+++ package.json
@@ -4,7 +4,6 @@
   "description": "Learn about Seattle neighborhoods",
   "main": "server.js",
   "dependencies": {
-    "cors": "^2.8.1",
     "ejs": "^2.5.5",
     "express": "^4.14.1",
     "eyes": "^0.1.8",
