--- routers/food/yelp_secrets.json
@@ -1,6 +1,0 @@
-{
-	"consumer_key": "Ubno_SkDE5Q9mftmU8LV6w",
-    "consumer_secret": "P6whm8VoDnD45OBipRNS-ONGgT8",
-    "token": "JTYXlT3BycqjVayPs2HnGDfejH5gZ5FI",
-    "token_secret": "GDBwiG2_fdESV5VjVhsKw5tvTIs"
-}
